refactor: move all spatial representation types to dedicated module
