+++ honeycomb-core/src/orbits.rs
@@ -216,7 +216,7 @@ mod tests {
     use crate::{CMap2, DartIdentifier, FloatType, Orbit2};
 
     fn simple_map() -> CMap2<FloatType> {
-        let mut map: CMap2<FloatType> = CMap2::new(6, 4);
+        let mut map: CMap2<FloatType> = CMap2::new(6);
         map.set_betas(1, [3, 2, 0]);
         map.set_betas(2, [1, 3, 4]);
         map.set_betas(3, [2, 1, 0]);
