+++ honeycomb-core/src/spatial_repr/vector.rs
